Add Compatibility with MultiResponse
ActiveMerchant::Billing::MultiResponse is an extension of ActiveMerchant::Billing::Response
that wraps multiple responses into one class. It does not set instance variables of the
response options, but rather defines getter methods to fetch these values from the primary response

If we want to store an instance of MultiResponse properly, we'll need to also check to see if the response
responds_to the field in question if it doesn't have an instance variable set for it.